Finished correct showing of expandlable item fro all options

+++ firstitemexpandablelistview/src/main/java/com/zo2m4bie/firstitemexpandablelistview/holder/ISelfExpandableHolder.java
@@ -5,6 +5,6 @@ package com.zo2m4bie.firstitemexpandablelistview.holder;
  */
 public interface ISelfExpandableHolder {
 
-    void heightPercentage(int percent);
+    void heightPercentage(float percent);
 
 }
